Remove eventCreation as a separate file
